requirements.txt
arabic-reshaper==2.1.4
asgiref==3.5.2
asn1crypto==1.5.1
autopep8==1.7.0
certifi==2022.9.24
cffi==1.15.1
charset-normalizer==2.1.1
click==8.1.3
cryptography==38.0.3
cssselect2==0.7.0
Django==4.1.2
docopt==0.6.2
future==0.18.2
html5lib==1.1
idna==3.4
lxml==4.9.1
oscrypto==1.3.0
Pillow==9.2.0
pipreqs==0.4.11
psycopg2-binary==2.9.4
pycodestyle==2.9.1
pycparser==2.21
pyHanko==0.15.1
pyhanko-certvalidator==0.19.6
PyJWT==2.6.0
PyPDF3==1.0.6
python-bidi==0.4.2
python-decouple==3.6
pytz==2022.5
pytz-deprecation-shim==0.1.0.post0
PyYAML==6.0
qrcode==7.3.1
reportlab==3.6.12
requests==2.28.1
six==1.16.0
sqlparse==0.4.3
svglib==1.4.1
tinycss2==1.2.1
toml==0.10.2
tqdm==4.64.1
twilio==7.15.0
tzdata==2022.6
tzlocal==4.2
uritools==4.0.0
urllib3==1.26.12
webencodings==0.5.1
xhtml2pdf==0.2.8
yarg==0.1.9
